Fix megaNuke not showing all banned users
Why does this group by the start timestamp, that does not make sense. A group by is an aggregation function, this basically is grouping all the things with the same start timestamp into a single record.

Pretty sure this fixes the mega nuke thing.

Either this, or changing the groupBy to groupBy u.username.