Improve readability of `subscriptionConfirm()`
As part of these improvements, this commit fixes a bug where an existing
subscription warning is only displayed when the existing sub is the same
tier as the sub being purchased. Considering the nature of the warning
(only the highest tier sub will be visible), it should be displayed in
any circumstance where a sub already exists.