Fix bug with updating stream end time
The cached datetime that marked the end of a stream was being converted
into a `bool` because of an incorrect type declaration for
`TwitchApiService->getStreamStatus()`. When this value was passed into
`Date::getDateTime()`, it threw an exception internally and simply
returned the current time. As a result, the stream end time was always
the current time.